Update to 1.13: First attempt (it compiles)
This is a massive update, chaning a ton of internals.
It broke a ton of obfuscated names, the entity registry system,
and ruined my nice API for spawn eggs.
It's been a while since , so I'm already jumping

Now that , I have to have a class .
I call but it's horribly broken because Material.valueOf doesn't work.
I fear this is some sort of bizarre classpath issue,
which I do **not** want to deal with right now.

I've put over 4.5 hours into this single commit and
I'm ready to call it a day.